Write smbget output to stdout
In 66c6a51 the `-O` flag to `smbget` was removed. On some platforms
(e.g. Ubuntu Noble; Debian 12.6) this flag means “socket options”. But
on other platforms (and in the Samba reference implementation) this flag
means “write output to stdout”.

As a result of the flag’s removal the call to `smbget` was downloading
the received JSON to a file in the working directory rather than writing
it to `stdout`. As a result, the `refresh()` function did not receive
the data and air quality values were zero value.

It appears the `--stdout` flag may be universally supported across
platforms to mean “write output to stdout” which should restore air
quality data regardless of platform.