Use smbclient and lower min. protocol to NT1
AirVisual Pro does not appear to support SMB protocol versions 2 or 3
which are the default versions supported by Samba circa December 2024.

Ubuntu Noble and Fedora 40 packages of Samba support an `--option`
flag which can be used with `smbget` to lower the supported protocol to
include NT1 without having to modify `smb.conf`.

Unfortunately, the Debian Samba package does not support this flag with
`smbget` and the flag does not appear in the official Samba man pages
for `smbget` on samba.org.

Fortunately, the `--option` flag _does_ appear to be supported on each
of Ubuntu Noble, Fedora 40, Debian 12.6 and in the official Samba
documentation for `smbclient`.

To resolve a silent failure to refresh data caused by a failed Samba
connection, rewrite the Samba command to use `smbclient` and lower the
minimum supported SMB protocol version to something AirVisual Pro can
speak.