Always set observations to latest received measurements
Otherwise, values like AQI may be incorrectly reported.

For example, the initial AQI value of the integration is initialized
to 0 and the initial AQI status is also initialized to 0 which evaluates
to an air quality status of “Unknown”.

Upon pulling the first set of measurements from the AirVisual Pro if
the observed AQI from the device is 0 this would not result in
re-evaluating the AQI status. With an observed AQI of 0 the AQI status
should now be reporting “Excellent”, however, rather than “Unknown”.

Instead, when new measurements are received they are always set and
evaluated even if the final result may be no change in the values.